Final_Project/requirements.txt
tensorflow>=2.10
keras
pillow
numpy
matplotlib
scikit-learn
opencv-python
streamlit
